python_server/requirements.txt
fastapi
uvicorn
pymongo
pydantic
dnspython
passlib 
cryptography 
openpyxl
python-jose
python-multipart
boto3
python-dotenv